Error handling for unsupported files and a few more tests
Sometimes Sharadar/Quandl returns a zero lenght dataframe for an
unsupported ticker. Other times it raises a ValueError. I now catch
both of these cases and continue with the other stocks.

Set the logging to INFO rather than the DEBUG.

Added a test to use the SF1 dataset as well as the SF0